+++ src/main/java/name/divinityunbound/block/custom/MobAttractorBlock.java
@@ -3,58 +3,63 @@ package name.divinityunbound.block.custom;
 import com.mojang.serialization.MapCodec;
 import name.divinityunbound.block.ModBlockConstants;
 import name.divinityunbound.block.ModBlocks;
-import name.divinityunbound.block.entity.DivineReplicatorBlockEntity;
 import name.divinityunbound.block.entity.MobAttractorBlockEntity;
 import name.divinityunbound.block.entity.ModBlockEntities;
-import name.divinityunbound.block.entity.SpaceSiphonBlockEntity;
+import name.divinityunbound.block.entity.UnholySilencerBlockEntity;
 import net.minecraft.block.*;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
+import net.minecraft.screen.NamedScreenHandlerFactory;
+import net.minecraft.state.StateManager;
+import net.minecraft.state.property.BooleanProperty;
+import net.minecraft.state.property.Properties;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.Hand;
 import net.minecraft.util.ItemScatterer;
+import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.random.Random;
-import net.minecraft.util.shape.VoxelShape;
-import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Iterator;
+import java.util.List;
 
-public class SpaceSiphonBlock extends BlockWithEntity implements BlockEntityProvider {
-    private static final VoxelShape SHAPE = Block.createCuboidShape(0, 0, 0, 16, 16, 16);
+public class MobAttractorBlock extends BlockWithEntity implements BlockEntityProvider {
     protected final Random random = Random.create();
-    public static final MapCodec<SpaceSiphonBlock> CODEC = SpaceSiphonBlock.createCodec(SpaceSiphonBlock::new);
-    public SpaceSiphonBlock(Settings settings) {
+    public static final BooleanProperty ENABLED = Properties.ENABLED;
+    public static final MapCodec<MobAttractorBlock> CODEC = MobAttractorBlock.createCodec(MobAttractorBlock::new);
+    public MobAttractorBlock(Settings settings) {
         super(settings);
+        this.setDefaultState(this.stateManager.getDefaultState().with(ENABLED, false));
     }
-
-    @Override
-    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        return SHAPE;
-    }
-
     @Override
     public BlockRenderType getRenderType(BlockState state) {
         return BlockRenderType.MODEL;
     }
-
     @Override
     protected MapCodec<? extends BlockWithEntity> getCodec() {
         return CODEC;
     }
 
+    @Override
+    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
+        builder.add(ENABLED);
+    }
     @Nullable
     @Override
     public BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
-        return new SpaceSiphonBlockEntity(pos, state);
+        return new MobAttractorBlockEntity(pos, state);
     }
 
     @Override
     public void neighborUpdate(BlockState state, World world, BlockPos pos, Block sourceBlock, BlockPos sourcePos, boolean notify) {
         if (!world.isClient) {
+            this.updateEnabled(world, pos, state);
             this.updateUpgrades(world, pos);
         }
     }
@@ -62,46 +67,52 @@ public class SpaceSiphonBlock extends BlockWithEntity implements BlockEntityProv
     @Override
     public void onPlaced(World world, BlockPos pos, BlockState state, LivingEntity placer, ItemStack itemStack) {
         if (!world.isClient) {
+            this.updateEnabled(world, pos, state);
             this.updateUpgrades(world, pos);
+        }
+    }
+
+    @Override
+    public void onBlockAdded(BlockState state, World world, BlockPos pos, BlockState oldState, boolean notify) {
+        if (oldState.isOf(state.getBlock())) {
+            return;
+        }
+        if (!world.isClient && world.getBlockEntity(pos) == null) {
+            this.updateEnabled(world, pos, state);
+        }
+    }
+
+    private void updateEnabled(World world, BlockPos pos, BlockState state) {
+        boolean bl = world.isReceivingRedstonePower(pos);
+        if (bl != state.get(ENABLED)) {
+            world.setBlockState(pos, (BlockState)state.with(ENABLED, bl), Block.NOTIFY_LISTENERS);
         }
     }
 
     private void updateUpgrades(World world, BlockPos pos) {
         BlockEntity be = world.getBlockEntity(pos);
-        if (be instanceof SpaceSiphonBlockEntity) {
-            ((SpaceSiphonBlockEntity)be).resetUpgrades();
+        if (be instanceof MobAttractorBlockEntity) {
+            ((MobAttractorBlockEntity)be).resetUpgrades();
             Iterator it = ModBlockConstants.UPGRADE_PROVIDER_OFFSETS.iterator();
             while (it.hasNext()) {
                 BlockPos blockPosOffset = (BlockPos) it.next();
                 BlockPos calcPos = new BlockPos(pos.getX() + blockPosOffset.getX(),
                         pos.getY() + blockPosOffset.getY(), pos.getZ() + blockPosOffset.getZ());
                 if (world.getBlockState(calcPos).getBlock().equals(ModBlocks.SPEED_UPGRADE)) {
-                    ((SpaceSiphonBlockEntity)be).increaseSpeed();
+                    ((MobAttractorBlockEntity)be).increaseSpeed();
                 } else if (world.getBlockState(calcPos).getBlock().equals(ModBlocks.QUANTITY_UPGRADE)) {
-                    ((SpaceSiphonBlockEntity)be).increaseQuantity();
+                    ((MobAttractorBlockEntity)be).increaseQuantity();
                 } else if (world.getBlockState(calcPos).getBlock().equals(ModBlocks.RANGE_UPGRADE)) {
-                    ((SpaceSiphonBlockEntity)be).increaseRange();
+                    ((MobAttractorBlockEntity)be).increaseRange();
                 }
             }
-        }
-    }
-
-    @Override
-    public void onStateReplaced(BlockState state, World world, BlockPos pos, BlockState newState, boolean moved) {
-        if (state.getBlock() != newState.getBlock()) {
-            BlockEntity blockEntity = world.getBlockEntity(pos);
-            if (blockEntity instanceof SpaceSiphonBlockEntity) {
-                ItemScatterer.spawn(world, pos, (SpaceSiphonBlockEntity)blockEntity);
-                world.updateComparators(pos,this);
-            }
-            super.onStateReplaced(state, world, pos, newState, moved);
         }
     }
 
     @Nullable
     @Override
     public <T extends BlockEntity> BlockEntityTicker<T> getTicker(World world, BlockState state, BlockEntityType<T> type) {
-        return validateTicker(type, ModBlockEntities.SPACE_SIPHON_BLOCK_ENTITY,
+        return validateTicker(type, ModBlockEntities.MOB_ATTRACTOR_BLOCK_ENTITY,
                 (world1, pos, state1, blockEntity) -> blockEntity.tick(world1, pos, state1));
     }
 }