+++ src/main/java/name/divinityunbound/block/entity/SpaceSiphonBlockEntity.java
@@ -46,8 +46,6 @@ public class SpaceSiphonBlockEntity extends BlockEntity implements ImplementedIn
     private static final int INPUT_SLOT = 0;
     private static final int DEFAULT_RANGE = 6;
 
-    private static final int CHECK_UPGRADE_TICKS = 20;
-
     private int speedCount = 0;
     private int quantityCount = 0;
     private int rangeCount = 0;
@@ -56,7 +54,6 @@ public class SpaceSiphonBlockEntity extends BlockEntity implements ImplementedIn
     private int progress = 0;
     private int maxProgress = 72;
     private int transferCooldown = 0;
-    private int upgradeCheck = 0;
 
     public SpaceSiphonBlockEntity(BlockPos pos, BlockState state) {
         super(ModBlockEntities.SPACE_SIPHON_BLOCK_ENTITY, pos, state);
@@ -95,6 +92,9 @@ public class SpaceSiphonBlockEntity extends BlockEntity implements ImplementedIn
         super.writeNbt(nbt);
         Inventories.writeNbt(nbt, inventory);
         nbt.putInt("space_siphon.progress", progress);
+        nbt.putInt("space_siphon.speedCount", speedCount);
+        nbt.putInt("space_siphon.quantityCount", quantityCount);
+        nbt.putInt("space_siphon.rangeCount", rangeCount);
     }
 
     @Override
@@ -102,6 +102,9 @@ public class SpaceSiphonBlockEntity extends BlockEntity implements ImplementedIn
         super.readNbt(nbt);
         Inventories.readNbt(nbt, inventory);
         progress = nbt.getInt("space_siphon.progress");
+        speedCount = nbt.getInt("space_siphon.speedCount");
+        quantityCount = nbt.getInt("space_siphon.quantityCount");
+        rangeCount = nbt.getInt("space_siphon.rangeCount");
     }
 
     public void tick(World world, BlockPos pos, BlockState state) {
@@ -124,11 +127,6 @@ public class SpaceSiphonBlockEntity extends BlockEntity implements ImplementedIn
                 (double) bp.getY() + 0.5, (double) bp.getZ() + 0.5);
         //pe.sendMessage(Text.literal("Stack nbt "));
         if (belowInventory != null) {
-            if (upgradeCheck >= CHECK_UPGRADE_TICKS) {
-                countUpgrades(world, pos);
-                upgradeCheck = 0;
-            }
-            upgradeCheck++;
             this.detectAndRetrieveItems(world, pos);
             ItemStack itemStack = this.getStack(INPUT_SLOT);
             if (!itemStack.isEmpty()) {
@@ -174,25 +172,22 @@ public class SpaceSiphonBlockEntity extends BlockEntity implements ImplementedIn
         this.progress = 0;
     }
 
-    public static final List<BlockPos> UPGRADE_PROVIDER_OFFSETS = BlockPos.stream(-1, 0, -1, 1, 0, 1).filter((pos) -> {
-        return Math.abs(pos.getX()) == 1 || Math.abs(pos.getZ()) == 1;
-    }).map(BlockPos::toImmutable).toList();
+    public void resetUpgrades() {
+        this.speedCount = 0;
+        this.quantityCount = 0;
+        this.rangeCount = 0;
+    }
 
-    public void countUpgrades(World world, BlockPos pos) {
-        speedCount = 0;
-        quantityCount = 0;
-        Iterator it = UPGRADE_PROVIDER_OFFSETS.iterator();
-        while(it.hasNext()) {
-            BlockPos blockPosOffset = (BlockPos)it.next();
-            BlockPos calcPos = new BlockPos(pos.getX() + blockPosOffset.getX(),
-                    pos.getY() + blockPosOffset.getY(), pos.getZ() + blockPosOffset.getZ());
-            if(world.getBlockState(calcPos).getBlock().equals(ModBlocks.SPEED_UPGRADE)) {
-                speedCount++;
-            }
-            else if(world.getBlockState(calcPos).getBlock().equals(ModBlocks.QUANTITY_UPGRADE)) {
-                quantityCount++;
-            }
-        }
+    public void increaseSpeed() {
+        this.speedCount++;
+    }
+
+    public void increaseQuantity() {
+        this.quantityCount++;
+    }
+
+    public void increaseRange() {
+        this.rangeCount++;
     }
 
     private boolean hasCraftingFinished() {
@@ -214,7 +209,7 @@ public class SpaceSiphonBlockEntity extends BlockEntity implements ImplementedIn
     }
 
     public void detectAndRetrieveItems(World world, BlockPos pos) {
-        int range = DEFAULT_RANGE / 2;
+        int range = (DEFAULT_RANGE + (rangeCount*2)) / 2;
         int posx = pos.getX();
         int posy = pos.getY();
         int posz = pos.getZ();